requirements.txt
Flask
Flask_SQLAlchemy
PyMySQL
python-dotenv
cryptography